fix tests, without experimental functions
