+++ app/src/androidTest/java/com/github/sdp/ratemyepfl/activity/RoomReviewActivityTest.kt
@@ -48,26 +48,4 @@ class RoomReviewActivityTest {
         onView(withId(R.id.id_room_info))
             .check(matches(withText(fakeRoom.toString())))
     }
-
-    /* This tests depends on the Review activity layout, so it doesn't work when launching
-       a fragment in isolation
-     */
-    @Test
-    fun nonNullArgumentsResetsAddReview() {
-        onView(withId(R.id.reviewNavigationView)).perform(CustomViewActions.navigateTo(R.id.addReviewFragment))
-        val comment = "Good"
-        val title = "Good title"
-        onView(withId(R.id.reviewRatingBar)).perform(
-            AddReviewFragmentTest.performSetRating(
-                ReviewRating.GOOD
-            )
-        )
-        onView(withId(R.id.addReviewComment)).perform(ViewActions.typeText(comment))
-        closeSoftKeyboard()
-        onView(withId(R.id.addReviewTitle)).perform(ViewActions.typeText(title))
-        closeSoftKeyboard()
-        onView(withId(R.id.doneButton)).perform(ViewActions.click())
-        onView(withId(R.id.addReviewComment)).check(matches(withText("")))
-        onView(withId(R.id.addReviewTitle)).check(matches(withText("")))
-    }
 }
